+++ src/db/migrations/20240624123748_critics.js
@@ -1,4 +1,6 @@
-exports.up = function(knex) {
+exports.up = async function(knex) {
+  const exists = await knex.schema.hasTable("critics");
+  if (!exists) {
     return knex.schema.createTable("critics", (table) => {
       table.increments("critic_id").primary();
       table.string("preferred_name").notNullable();
@@ -6,9 +8,12 @@ exports.up = function(knex) {
       table.string("organization_name").notNullable();
       table.timestamps(true, true);
     });
-  };
-  
-  exports.down = function(knex) {
+  }
+};
+
+exports.down = async function(knex) {
+  const exists = await knex.schema.hasTable("critics");
+  if (exists) {
     return knex.schema.dropTable("critics");
-  };
-  
+  }
+};
