+++ src/db/migrations/20240624123756_reviews.js
@@ -1,4 +1,6 @@
-exports.up = function(knex) {
+exports.up = async function(knex) {
+  const exists = await knex.schema.hasTable("reviews");
+  if (!exists) {
     return knex.schema.createTable("reviews", (table) => {
       table.increments("review_id").primary();
       table.text("content").notNullable();
@@ -7,9 +9,12 @@ exports.up = function(knex) {
       table.integer("movie_id").unsigned().notNullable().references("movie_id").inTable("movies").onDelete("CASCADE");
       table.timestamps(true, true);
     });
-  };
-  
-  exports.down = function(knex) {
+  }
+};
+
+exports.down = async function(knex) {
+  const exists = await knex.schema.hasTable("reviews");
+  if (exists) {
     return knex.schema.dropTable("reviews");
-  };
-  
+  }
+};
