+++ src/db/migrations/20240624123810_movie_theaters.js
@@ -1,13 +1,18 @@
-exports.up = function(knex) {
+exports.up = async function(knex) {
+  const exists = await knex.schema.hasTable("movies_theaters");
+  if (!exists) {
     return knex.schema.createTable("movies_theaters", (table) => {
       table.integer("movie_id").unsigned().notNullable().references("movie_id").inTable("movies").onDelete("CASCADE");
       table.integer("theater_id").unsigned().notNullable().references("theater_id").inTable("theaters").onDelete("CASCADE");
       table.boolean("is_showing").defaultTo(false);
       table.timestamps(true, true);
     });
-  };
-  
-  exports.down = function(knex) {
+  }
+};
+
+exports.down = async function(knex) {
+  const exists = await knex.schema.hasTable("movies_theaters");
+  if (exists) {
     return knex.schema.dropTable("movies_theaters");
-  };
-  
+  }
+};
